refactor: mutation 훅 분리
