+++ frontend/src/pages/detail/hooks/useHighlights.ts
@@ -1,0 +1,13 @@
+import { useQuery } from '@tanstack/react-query';
+import { queries } from '@/apis/queries';
+
+export const useHighlights = ({ articleId }: { articleId: number }) => {
+  const { data: highlights, isSuccess } = useQuery(
+    queries.highlights({ articleId }),
+  );
+
+  return {
+    highlights: highlights?.content ?? [],
+    isHighlightLoaded: isSuccess,
+  };
+};
