+++ frontend/src/pages/detail/hooks/useUpdateHighlightMutation.ts
@@ -1,0 +1,16 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { patchHighlight, PatchHighlightParams } from '@/apis/highlight';
+
+export const useUpdateHighlightMutation = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: (params: PatchHighlightParams) => patchHighlight(params),
+
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ['highlight'],
+      });
+    },
+  });
+};
